+++ CHANGELOG.md
@@ -1,14 +1,17 @@
 Percona Monitoring and Management (PMM) Client
 
-v1.0.7 unreleased 2016-11-10
+v1.0.7 unreleased 2016-11-21
 
 * Added --bind-address flag to support running PMM server and client on the different networks.
   By default, this address is the same as client one. When running PMM on different networks, --client-address should be set to remote (public) address
   and --bind-address to local (private) address. This also assumes you configure NAT and port forwarding between those addresses.
-* Amended output of systemv service status if run adhoc (requires re-adding services).
+* Enable http basic auth for metric services automatically when it is configured on PMM server (applies to services added after the password was set).
+* MongoDB URI is not passed as command-line argument now and hidden from process list (requires re-adding mongodb:metrics service).
 * Expose PXC/Galera gcache size as a metric.
+* Use terminal color instead of emoji on "check-network" output and also "list" one.
+* Amended output of systemv service status if run adhoc (requires re-adding services).
 
-v1.0.6 unreleased 2016-11-04
+v1.0.6 released 2016-11-15
 
 * Fixes for "mysql:queries" service using perfschema query source:
   * do not crash when DIGEST_TEXT is NULL